disabling Github calls for now

+++ cookbooks/apitrary_pytools/recipes/balancr.rb
@@ -20,6 +20,7 @@ end
 execute "pip-install-deployr" do
   command "pip install git+ssh://git@github.com/apitrary/deployr.git"
   user "root"
+  not_if {File.exists?("/usr/local/bin/deployr.py")}
 end
 
 directory "/etc/deployr" do
